requirements.txt
Flask==2.0.2
requests==2.26.0
Jinja2==3.0.1
plotly==5.3.1
pandas==1.2.3
beautifulsoup4==4.10.0
dash==2.0.0
python-dotenv==0.19.2
PyYAML==6.0
shikimori_api==1.0.1
